BackEnd/requirements.txt
boto3
Flask
python-dotenv
opencv-python
numpy
json
flask_cors